Address issue #814 (#1072)
When the `showMousePosition` option is set to `true` and the mouse pointer moves over a non-2d track, x and y values are published internally that can cause another 2d track to draw nonsensical vertical or horizontal lines. This test clears the relevant value so that it cannot be used to render such lines.